+++ eultheme/context_processors.py
@@ -1,5 +1,6 @@
 from django.conf import settings
-
+from django.contrib.sites.models import get_current_site
+from django.utils.functional import SimpleLazyObject
 
 def template_settings(request):
     '''Template context processor: add selected setting to context
@@ -11,3 +12,15 @@ def template_settings(request):
     }
 
     return context_extras
+
+def site_path(request):
+    '''Template context processor: provides access to current
+    :class:`~django.contrib.sites.models.Site` and
+    the site root path for use in building absolute urls.
+    '''
+    site = SimpleLazyObject(lambda: get_current_site(request))
+    protocol = 'https' if request.is_secure() else 'http'
+    return {
+        'site': site,
+        'site_root': SimpleLazyObject(lambda: "%s://%s" % (protocol, site.domain))
+    }